devlink: Expose port function commands to control header split
Expose port function commands to enable / disable header split.

When header split is enabled the function is configured to place frame
headers and data into separate buffers. This is a pre-requisite for
TCP zero-copy receive.

The intention is to provide a simple toggle for this capability,
allowing the host to control the availability of this feature to
be configured by the driver.

Example of a PCI VF port which supports header split capability configuration:

    $ devlink port show pci/0000:06:00.0/2
    pci/0000:06:00.0/2: type eth netdev enp6s0pf0vf1 flavour pcivf pfnum 0 vfnum 1
        function:
            hw_addr 00:00:00:00:00:00 header_split enable

    $ devlink port function set pci/0000:06:00.0/2 header_split disable

    $ devlink port show pci/0000:06:00.0/2
    pci/0000:06:00.0/2: type eth netdev enp6s0pf0vf1 flavour pcivf pfnum 0 vfnum 1
        function:
            hw_addr 00:00:00:00:00:00 header_split disable

Based in part on:
-  commit 390a24cbc396 ("devlink: Expose port function commands to control IPsec packet offloads")
   by Dima Chumak <dchumak@nvidia.com>
-  commit 9690ae604290 ("ethtool: add header/data split indication")
   by Jakub Kicinski <kuba@kernel.org>

Signed-off-by: Simon Horman <horms@kernel.org>