devlink: Expose port function commands to control hardware GRO
Expose port function commands to enable / disable hardware GRO.

When hardware GRO is enabled the NIC will coalesce packets and pass them as
a coalesced units to the driver. The coalesced packet stream is
re-segmentable by GSO or TSO back to the exact original packet stream.

The intention is to provide a simple toggle for this capability,
allowing the host to control the availability of this feature to
be configured by the driver.

Example of a PCI VF port which supports hardware GRO split capability
configuration:

    $ devlink port show pci/0000:06:00.0/2
    pci/0000:06:00.0/2: type eth netdev enp6s0pf0vf1 flavour pcivf pfnum 0 vfnum 1
        function:
            hw_addr 00:00:00:00:00:00 rx-gro-hw enable

    $ devlink port function set pci/0000:06:00.0/2 rx-gro-hw disable

    $ devlink port show pci/0000:06:00.0/2
    pci/0000:06:00.0/2: type eth netdev enp6s0pf0vf1 flavour pcivf pfnum 0 vfnum 1
        function:
            hw_addr 00:00:00:00:00:00 rx-gro-hw disable

Signed-off-by: Simon Horman <horms@kernel.org>
--
v0.2
- Correct grammar documentation of ethtool in devlink-port.rst
- Don't remove ' *' line from Kernel doc for devlink_port_ops
v0.1
- Corrected copy-paste error so that documentation always refers to gro
- Updated name of option from gro to rx-gro-hw, which is already
  used for the related ethtool option. Made this naming consistent
  throughout.

Signed-off-by: Simon Horman <horms@kernel.org>